typer skal ligge i egen types fil. Forenkler metode for initialisering av valgte ve inntektstyper
